Make safe for Int32
10^15 = 0 for Int32, so this evaluated to `Inf` and thus was always true on Windows 32. Not a huge deal, but this mean that the initial dt would take the wrong path on Windows 32 sometimes (actually, most of the time it wouldn't, but in the cases where the equation was too stiff at the start it would... really weird...). This manually up-converts to Int64 since Int64(10)^(15) is sufficiently large enough to hold the rational correctly, and let it convert to whatever the type of t is just fine. The other way this could've been handled is `1e-15`, but there's essentially no performance difference and this will keep the ability for exact representation in whatever user number type.

+++ src/initdt.jl
@@ -48,7 +48,7 @@ function ode_determine_initdt{tType,uType}(u0,t::tType,tdir,dtmax,abstol,reltol,
   d₂ = internalnorm(tmp)/dt₀
   # Hairer has d₂ = sqrt(sum(abs2,tmp))/dt₀, note the lack of norm correction
   unitless_max = max(d₁/typeof(d₁)(one(d₁)),d₂/typeof(d₂)(one(d₂)))
-  if unitless_max<=1//10^(15)
+  if unitless_max<=1//Int64(10)^(15)
     dt₁ = max(tType(1//10^(6)),dt₀*1//10^(3))
   else
     dt₁ = tType(10.0^(-(2+log10(unitless_max))/(order)))
@@ -76,7 +76,7 @@ function ode_determine_initdt{uType,tType}(u0::uType,t,tdir,dtmax,abstol,reltol,
   u₁ = u0 + tdir*dt₀*f₀
   f₁ = f(t+tdir*dt₀,u₁)
   d₂ = internalnorm((f₁-f₀)./(abstol+abs.(u0).*reltol))/dt₀*tType(1)
-  if max(d₁,d₂) <= T1(1//10^(15))
+  if max(d₁,d₂) <= T1(1//Int64(10)^(15))
     dt₁ = max(tType(1//10^(6)),dt₀*1//10^(3))
   else
     dt₁ = tType(10.0^(-(2+log10(max(d₁,d₂)/T1(1)))/(order)))
